requirements.txt
PyQt5==5.15.9
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2 